backend/requirements.txt
# Framework webowy
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10

# ORM i baza danych
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.21
pymysql==1.1.0
cryptography==41.0.4  # Wymagane dla PyMySQL z SSL

# Biblioteki astrologiczne
pyswisseph==2.10.3    # Swiss Ephemeris - precyzyjne obliczenia planet
jyotishyayoga==0.2.7  # Obliczanie joga wedyjskich
flatlib==0.3.1        # Biblioteka do astrologii

# Integracja z OpenAI
openai==1.3.0

# Narzędzia do geolokalizacji i obliczeń
pytz==2023.3
timezonefinder==6.2.0
geopy==2.4.0

# Testy i weryfikacja
pytest==7.4.2
pytest-flask==1.2.0

# Inne
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
